ui: complete v1 solargen dashboard

+++ solargen-ui/src/hooks/useTime.ts
@@ -1,17 +1,20 @@
 import {useEffect, useState} from 'react';
 
 export function useTime() {
+    // Hooks
     const [time, setTime] = useState(new Date());
 
+    // Effects
     useEffect(() => {
         const interval = setInterval(() => setTime(new Date()), 1000);
         return () => clearInterval(interval);
     }, []);
 
+    // Misc.
     const currentDate = time.toLocaleDateString('fr-FR', {
         timeZone: 'Australia/Melbourne',
         day: 'numeric',
-        month: 'long',
+        month: 'short',
         year: 'numeric',
     });
